requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
asyncdagpi==3.2.2
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
discord==1.0.1
discord-pretty-help==1.2.0
discord.py==1.6.0
distlib==0.3.1
Django==3.1.4
dnspython==2.1.0
dpymenus==1.2.3
email-verifier==1.0.0
emoji==0.6.0
filelock==3.0.12
Flask==1.1.2
Flask-EmailVerifier==0.1.0
future==0.18.2
google==3.0.0
idna==2.10
image==1.5.33
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
ksoftapi==0.2.5a0
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mee6-py-api==0.0.1
multidict==4.7.6
mysql-connector-python==8.0.22
nekobot==1.2
nose==1.3.7
numpy==1.19.5
pandas==1.2.1
pickle-mixin==1.0.2
Pillow==8.0.1
pip==21.0.1
praw==7.1.0
prawcore==1.5.0
prettyprint==0.1.5
protobuf==3.14.0
pycparser==2.20
pylint==2.6.0
pymongo==3.11.3
PyMySQL==0.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
selenium==3.141.0
setuptools==51.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse ==0.4.1
stripe==2.55.0
TenGiphPy==2.0.0
toml==0.10.2
udpy==2.0.0
update-checker==0.18.0
urllib3==1.25.11
virtualenv==20.1.0
websocket-client==0.57.0
Werkzeug==1.0.1
wheel==0.35.1
wrapt==1.12.1
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.5.1